chore: move supabase.js file
